split revurdering og søknad fra sakIntro & legg til feilmeldinger
