+++ src/components/apiErrorAlert/revurderingApiError/RevurderingApiError.ts
@@ -29,7 +29,8 @@ export type RevurderingErrorCodes =
     | UtfallSomIkkeStøttes
     | Brev
     | Stans
-    | Gjenopptak;
+    | Gjenopptak
+    | Avsluttet;
 
 type VilkårErrorCodes = OpprettelseOgOppdatering | Uføre | Bosituasjon | Formue | Fradrag;
 
@@ -45,8 +46,6 @@ enum Generell {
     UGYLDIG_ÅRSAK = 'ugyldig_årsak',
 
     KUNNE_IKKE_SLÅ_OPP_EPS = 'kunne_ikke_slå_opp_eps',
-
-    REVURDERINGEN_ER_ALLEREDE_AVSLUTTET = 'revurderingen_er_allerede_avsluttet',
 }
 
 enum Vurderingsperiode {
@@ -149,6 +148,13 @@ enum Gjenopptak {
     IVERKSETTING_FØRER_TIL_FEILUTBETALING = 'kunne_ikke_iverksette_gjenopptak_fører_til_feilutbetaling',
 }
 
+enum Avsluttet {
+    REVURDERINGEN_ER_ALLEREDE_AVSLUTTET = 'revurderingen_er_allerede_avsluttet',
+    REVURDERINGER_ER_TIL_ATTESTERING = ' ',
+    REVURDERINGEN_ER_IVERKSATT = 'revurderingen_er_iverksatt',
+    REVURDERING_ER_IKKE_FORHÅNDSVARSLET_FOR_Å_VISE_BREV = 'revurdering_er_ikke_forhåndsvarslet_for_å_vise_brev',
+}
+
 const revurderingErrorCodeMessageIdMap: { [key in RevurderingErrorCodes]: keyof typeof messages | undefined } = {
     [Generell.G_REGULERING_KAN_IKKE_FØRE_TIL_OPPHØR]: 'generell.gregulering.kan.ikke.føre.til.opphør',
     [Generell.ATTESTANT_OG_SAKSBEHANDLER_KAN_IKKE_VÆRE_SAMME_PERSON]:
@@ -159,7 +165,6 @@ const revurderingErrorCodeMessageIdMap: { [key in RevurderingErrorCodes]: keyof 
     [Generell.UGYLDIG_PERIODE]: 'generell.ugyldig.periode',
     [Generell.UGYLDIG_ÅRSAK]: 'generell.ugyldig.årsak',
     [Generell.KUNNE_IKKE_SLÅ_OPP_EPS]: 'generell.kunne.ikke.slå.opp.eps',
-    [Generell.REVURDERINGEN_ER_ALLEREDE_AVSLUTTET]: 'generell.revurderingen.er.allerede.avsluttet',
 
     [Vurderingsperiode.INGENTING_Å_REVURDERE_I_PERIODEN]: 'periode.ingenting.å.revurdere',
 
@@ -243,4 +248,10 @@ const revurderingErrorCodeMessageIdMap: { [key in RevurderingErrorCodes]: keyof 
     [Gjenopptak.KAN_IKKE_GJENOPPTA_OPPHØRTE_UTBETALINGER]: 'gjenopptak.kan.ikke.gjenoppta.opphørte.utbetalinger',
     [Gjenopptak.ÅPEN_REVURDERING_EKSISTERER]: 'gjenopptak.åpen.revurdering.eksisterer',
     [Gjenopptak.IVERKSETTING_FØRER_TIL_FEILUTBETALING]: 'gjenopptak.iverksetting.feilutbetaling',
+
+    [Avsluttet.REVURDERINGEN_ER_ALLEREDE_AVSLUTTET]: 'avsluttet.revurderingen.er.allerede.avsluttet',
+    [Avsluttet.REVURDERINGER_ER_TIL_ATTESTERING]: 'avsluttet.revurderingen.er.til.attestering',
+    [Avsluttet.REVURDERINGEN_ER_IVERKSATT]: 'avsluttet.revurderingen.er.iverksatt',
+    [Avsluttet.REVURDERING_ER_IKKE_FORHÅNDSVARSLET_FOR_Å_VISE_BREV]:
+        'avsluttet.revurderingen.er.ikke.forhåndsvarslet.for.å.vise.brev',
 };
