+++ src/pages/saksbehandling/sakintro/Sakintro.tsx
@@ -1,52 +1,31 @@
 import * as RemoteData from '@devexperts/remote-data-ts';
-import { Alert, BodyShort, Button, Heading, Label, LinkPanel, Loader, Panel, Tag } from '@navikt/ds-react';
-import { isEmpty, last } from 'fp-ts/lib/Array';
-import { toNullable } from 'fp-ts/lib/Option';
-import Ikon from 'nav-frontend-ikoner-assets';
+import { Alert, Button, Heading, LinkPanel, Loader } from '@navikt/ds-react';
+import { isEmpty } from 'fp-ts/lib/Array';
 import React from 'react';
 import { IntlShape } from 'react-intl';
-import { Link, useHistory } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
 import { FeatureToggle } from '~api/featureToggleApi';
 import { ÅpentBrev } from '~assets/Illustrations';
-import ApiErrorAlert from '~components/apiErrorAlert/ApiErrorAlert';
 import LinkAsButton from '~components/linkAsButton/LinkAsButton';
-import UnderkjenteAttesteringer from '~components/underkjenteAttesteringer/UnderkjenteAttesteringer';
-import { useUserContext } from '~context/userContext';
-import * as sakSlice from '~features/saksoversikt/sak.slice';
 import { useFeatureToggle } from '~lib/featureToggles';
-import { pipe } from '~lib/fp';
-import { ApiResult, useAsyncActionCreator, useNotificationFromLocation } from '~lib/hooks';
+import { ApiResult, useNotificationFromLocation } from '~lib/hooks';
 import { useI18n } from '~lib/i18n';
 import * as Routes from '~lib/routes';
 import { Nullable } from '~lib/types';
 import Utbetalinger from '~pages/saksbehandling/sakintro/Utbetalinger';
 import { Behandling } from '~types/Behandling';
-import { AbstraktRevurdering, RevurderingsStatus } from '~types/Revurdering';
 import { Sak } from '~types/Sak';
-import { LukkSøknadBegrunnelse, Søknad } from '~types/Søknad';
-import { Vedtak } from '~types/Vedtak';
-import { erIverksatt, erTilAttestering, hentSisteVurdertSaksbehandlingssteg } from '~utils/behandling/behandlingUtils';
-import {
-    getIverksatteInnvilgedeSøknader,
-    søknadMottatt,
-    getIverksatteAvslåtteSøknader,
-} from '~utils/søknad/søknadUtils';
+import { Søknad } from '~types/Søknad';
+import { erIverksatt } from '~utils/behandling/behandlingUtils';
+import { getIverksatteInnvilgedeSøknader, getIverksatteAvslåtteSøknader } from '~utils/søknad/søknadUtils';
 
-import {
-    erRevurderingTilAttestering,
-    erRevurderingIverksatt,
-    erRevurderingSimulert,
-    erForhåndsvarselSendt,
-    finnNesteRevurderingsteg,
-    erRevurderingStans,
-    erRevurderingGjenopptak,
-    erRevurdering,
-} from '../../../utils/revurdering/revurderingUtils';
-import { RevurderingSteg } from '../types';
+import { getAvsluttedeOgIkkeAvsluttedeRevurderinger } from '../../../utils/revurdering/revurderingUtils';
 
+import { AvsluttedeRevurderinger, ÅpneRevurderinger } from './RevurderingsLister';
 import messages from './sakintro-nb';
 import styles from './sakintro.module.less';
+import { AvslåtteSøknader, IverksattInnvilgedeSøknader, LukkedeSøknader, ÅpneSøknader } from './SøknadsLister';
 
 const SuksessStatuser = (props: { locationState: Nullable<Routes.SuccessNotificationState> }) => {
     return (
@@ -54,19 +33,6 @@ const SuksessStatuser = (props: { locationState: Nullable<Routes.SuccessNotifica
             {props.locationState?.notification && <Alert variant="success">{props.locationState.notification}</Alert>}
         </div>
     );
-};
-
-const lukketBegrunnelseResourceId = (type?: LukkSøknadBegrunnelse) => {
-    switch (type) {
-        case LukkSøknadBegrunnelse.Avvist:
-            return 'søknad.lukket.avvist';
-        case LukkSøknadBegrunnelse.Bortfalt:
-            return 'søknad.lukket.bortfalt';
-        case LukkSøknadBegrunnelse.Trukket:
-            return 'søknad.lukket.trukket';
-        default:
-            return 'søknad.lukket.ukjentLukking';
-    }
 };
 
 const Sakintro = (props: { sak: Sak }) => {
@@ -90,7 +56,10 @@ const Sakintro = (props: { sak: Sak }) => {
 
     const avslåtteSøknader = getIverksatteAvslåtteSøknader(props.sak);
 
-    const revurderinger = props.sak.revurderinger;
+    const { avsluttedeRevurderinger, åpneRevurderinger } = getAvsluttedeOgIkkeAvsluttedeRevurderinger(
+        props.sak.revurderinger
+    );
+
     const kanRevurderes = !isEmpty(props.sak.utbetalinger);
 
     const revurderingToggle = useFeatureToggle(FeatureToggle.Revurdering) && kanRevurderes;
@@ -129,7 +98,9 @@ const Sakintro = (props: { sak: Sak }) => {
                         utbetalingsperioder={props.sak.utbetalinger}
                         kanStansesEllerGjenopptas={props.sak.utbetalingerKanStansesEllerGjenopptas}
                     />
-                    {revurderingToggle && <Revurderinger sak={props.sak} revurderinger={revurderinger} intl={intl} />}
+                    {revurderingToggle && (
+                        <ÅpneRevurderinger sak={props.sak} åpneRevurderinger={åpneRevurderinger} intl={intl} />
+                    )}
                     <IverksattInnvilgedeSøknader
                         sak={props.sak}
                         iverksatteInnvilgedeSøknader={iverksatteInnvilgedeSøknader}
@@ -137,6 +108,7 @@ const Sakintro = (props: { sak: Sak }) => {
                     />
                     <AvslåtteSøknader sak={props.sak} avslåtteSøknader={avslåtteSøknader} intl={intl} />
                     <LukkedeSøknader lukkedeSøknader={lukkedeSøknader} intl={intl} />
+                    <AvsluttedeRevurderinger avsluttedeRevurderinger={avsluttedeRevurderinger} intl={intl} />
                     <div>
                         <LinkPanel
                             href={Routes.alleDokumenterForSak.createURL({ sakId: props.sak.id })}
@@ -159,535 +131,7 @@ const Sakintro = (props: { sak: Sak }) => {
     );
 };
 
-const ÅpneSøknader = (props: {
-    åpneSøknader: Søknad[];
-    behandlinger: Behandling[];
-    sakId: string;
-    intl: IntlShape;
-}) => {
-    if (props.åpneSøknader.length === 0) return null;
-
-    return (
-        <div className={styles.søknadsContainer}>
-            <Heading level="2" size="medium" spacing>
-                {props.intl.formatMessage({ id: 'søknad.åpneSøknader.tittel' })}
-            </Heading>
-            <ol>
-                {props.åpneSøknader.map((s) => {
-                    const behandling = props.behandlinger.find((b) => b.søknad.id === s.id);
-                    const attesteringer = behandling?.attesteringer ?? [];
-                    const senesteAttestering = pipe(attesteringer, last, toNullable);
-
-                    return (
-                        <div key={s.id}>
-                            <Panel border className={styles.søknad}>
-                                <div className={styles.info}>
-                                    <div>
-                                        <Heading level="3" size="medium">
-                                            {props.intl.formatMessage({ id: 'søknad.typeSøknad' })}
-                                        </Heading>
-                                        <div className={styles.dato}>
-                                            <Label>{`${props.intl.formatMessage({ id: 'søknad.mottatt' })}: `}</Label>
-                                            <BodyShort>{søknadMottatt(s, props.intl)}</BodyShort>
-                                        </div>
-                                        {senesteAttestering?.underkjennelse && (
-                                            <UnderkjenteAttesteringer attesteringer={attesteringer} />
-                                        )}
-                                    </div>
-                                    <div className={styles.knapper}>
-                                        {!behandling ? (
-                                            <StartSøknadsbehandlingKnapper
-                                                sakId={props.sakId}
-                                                søknadId={s.id}
-                                                intl={props.intl}
-                                            />
-                                        ) : (
-                                            <SøknadsbehandlingStartetKnapper
-                                                sakId={props.sakId}
-                                                intl={props.intl}
-                                                behandling={behandling}
-                                                søknadId={s.id}
-                                            />
-                                        )}
-                                    </div>
-                                </div>
-                            </Panel>
-                        </div>
-                    );
-                })}
-            </ol>
-        </div>
-    );
-};
-
-const Revurderinger = (props: { sak: Sak; revurderinger: AbstraktRevurdering[]; intl: IntlShape }) => {
-    if (props.revurderinger.length === 0) return null;
-
-    return (
-        <div className={styles.søknadsContainer}>
-            <Heading level="2" size="medium" spacing>
-                {props.intl.formatMessage({ id: 'revurdering.tittel' })}
-            </Heading>
-            <ol>
-                {props.revurderinger.map((r) => {
-                    const vedtakForBehandling = props.sak.vedtak.find((v) => v.behandlingId === r.id);
-                    const underkjenteRevurderinger = r.attesteringer.filter((a) => a.underkjennelse !== null);
-                    return (
-                        <div key={r.id}>
-                            <Panel border className={styles.søknad}>
-                                <div className={styles.info}>
-                                    <div>
-                                        <div className={styles.tittel}>
-                                            <Heading level="3" size="small" spacing>
-                                                {props.intl.formatMessage({ id: 'revurdering.undertittel' })}
-                                            </Heading>
-                                            {erRevurdering(r) && erForhåndsvarselSendt(r) && (
-                                                <Tag variant="info" className={styles.etikett}>
-                                                    {props.intl.formatMessage({
-                                                        id: 'revurdering.label.forhåndsvarselSendt',
-                                                    })}
-                                                </Tag>
-                                            )}
-                                        </div>
-                                        <div className={styles.dato}>
-                                            <Label>{props.intl.formatMessage({ id: 'revurdering.opprettet' })} </Label>
-                                            <BodyShort>{props.intl.formatDate(r.opprettet)}</BodyShort>
-                                        </div>
-                                        {vedtakForBehandling?.opprettet && (
-                                            <div className={styles.dato}>
-                                                <Label>
-                                                    {props.intl.formatMessage({ id: 'revurdering.iverksattDato' })}{' '}
-                                                </Label>
-                                                <BodyShort>
-                                                    {props.intl.formatDate(vedtakForBehandling.opprettet)}
-                                                </BodyShort>
-                                            </div>
-                                        )}
-                                        {underkjenteRevurderinger.length > 0 &&
-                                            erRevurdering(r) &&
-                                            !erRevurderingIverksatt(r) && (
-                                                <div className={styles.underkjenteAttesteringerContainer}>
-                                                    <UnderkjenteAttesteringer attesteringer={r.attesteringer} />
-                                                </div>
-                                            )}
-                                    </div>
-                                    <div className={styles.knapper}>
-                                        <RevurderingStartetKnapper
-                                            sakId={props.sak.id}
-                                            vedtak={props.sak.vedtak}
-                                            revurdering={r}
-                                            intl={props.intl}
-                                        />
-                                    </div>
-                                </div>
-                            </Panel>
-                        </div>
-                    );
-                })}
-            </ol>
-        </div>
-    );
-};
-
-const RevurderingStartetKnapper = (props: {
-    sakId: string;
-    revurdering: AbstraktRevurdering;
-    vedtak: Vedtak[];
-    intl: IntlShape;
-}) => {
-    const user = useUserContext();
-    const { revurdering } = props;
-    const vedtak = props.vedtak.find((v) => v.behandlingId === revurdering.id);
-
-    return (
-        <div className={styles.behandlingContainer}>
-            {erRevurdering(revurdering) &&
-                erRevurderingTilAttestering(revurdering) &&
-                (!user.isAttestant || user.navIdent === revurdering.saksbehandler) && (
-                    <div className={styles.ikonContainer}>
-                        <Ikon className={styles.ikon} kind="info-sirkel-fyll" width={'24px'} />
-                        <p>
-                            {props.intl.formatMessage({
-                                id: 'attestering.tilAttestering',
-                            })}
-                        </p>
-                    </div>
-                )}
-
-            {erRevurdering(revurdering) && erRevurderingIverksatt(revurdering) && vedtak && (
-                <LinkAsButton
-                    variant="secondary"
-                    href={Routes.vedtaksoppsummering.createURL({ sakId: props.sakId, vedtakId: vedtak.id })}
-                    size="small"
-                >
-                    Se oppsummering
-                </LinkAsButton>
-            )}
-
-            <div className={styles.knapper}>
-                {erRevurdering(revurdering) &&
-                erRevurderingTilAttestering(revurdering) &&
-                user.isAttestant &&
-                user.navIdent !== revurdering.saksbehandler ? (
-                    <LinkAsButton
-                        variant="secondary"
-                        size="small"
-                        href={Routes.attesterRevurdering.createURL({
-                            sakId: props.sakId,
-                            revurderingId: revurdering.id,
-                        })}
-                    >
-                        {props.intl.formatMessage({
-                            id: 'attestering.attester',
-                        })}
-                    </LinkAsButton>
-                ) : erRevurderingStans(revurdering) ? (
-                    <AvsluttOgStartFortsettButtons
-                        sakId={props.sakId}
-                        behandlingsId={revurdering.id}
-                        primaryButtonTekst={
-                            revurdering.status === RevurderingsStatus.IVERKSATT_STANS
-                                ? props.intl.formatMessage({ id: 'revurdering.oppsummering' })
-                                : props.intl.formatMessage({ id: 'revurdering.fortsett' })
-                        }
-                        usePrimaryAsLink={{
-                            url: Routes.stansOppsummeringRoute.createURL({
-                                sakId: props.sakId,
-                                revurderingId: revurdering.id,
-                            }),
-                        }}
-                        intl={props.intl}
-                    />
-                ) : erRevurderingGjenopptak(revurdering) ? (
-                    <AvsluttOgStartFortsettButtons
-                        sakId={props.sakId}
-                        behandlingsId={revurdering.id}
-                        primaryButtonTekst={
-                            revurdering.status === RevurderingsStatus.IVERKSATT_GJENOPPTAK
-                                ? props.intl.formatMessage({ id: 'revurdering.oppsummering' })
-                                : props.intl.formatMessage({ id: 'revurdering.fortsett' })
-                        }
-                        usePrimaryAsLink={{
-                            url: Routes.gjenopptaStansOppsummeringRoute.createURL({
-                                sakId: props.sakId,
-                                revurderingId: revurdering.id,
-                            }),
-                        }}
-                        intl={props.intl}
-                    />
-                ) : (
-                    erRevurdering(revurdering) &&
-                    !erRevurderingTilAttestering(revurdering) &&
-                    !erRevurderingIverksatt(revurdering) &&
-                    user.navIdent !== pipe(revurdering.attesteringer, last, toNullable)?.attestant && (
-                        <AvsluttOgStartFortsettButtons
-                            sakId={props.sakId}
-                            behandlingsId={revurdering.id}
-                            primaryButtonTekst={props.intl.formatMessage({ id: 'revurdering.fortsett' })}
-                            usePrimaryAsLink={{
-                                url: Routes.revurderValgtRevurdering.createURL({
-                                    sakId: props.sakId,
-                                    steg: erRevurderingSimulert(revurdering)
-                                        ? RevurderingSteg.Oppsummering
-                                        : finnNesteRevurderingsteg(revurdering.informasjonSomRevurderes),
-                                    revurderingId: revurdering.id,
-                                }),
-                            }}
-                            intl={props.intl}
-                        />
-                    )
-                )}
-            </div>
-        </div>
-    );
-};
-
-const IverksattInnvilgedeSøknader = (props: {
-    iverksatteInnvilgedeSøknader: Array<{
-        iverksattDato: string | undefined;
-        søknadensBehandlingsId: string | undefined;
-        søknad: Søknad;
-    }>;
-    sak: Sak;
-    intl: IntlShape;
-}) => {
-    if (props.iverksatteInnvilgedeSøknader.length === 0) return null;
-
-    return (
-        <div className={styles.søknadsContainer}>
-            <Heading level="2" size="medium" spacing>
-                {props.intl.formatMessage({ id: 'søknad.godkjenteSøknader.tittel' })}
-            </Heading>
-            <ol>
-                {props.iverksatteInnvilgedeSøknader.map((s) => {
-                    if (!s.søknadensBehandlingsId) {
-                        return null;
-                    }
-                    const vedtak = props.sak.vedtak.find((v) => v.behandlingId === s.søknadensBehandlingsId);
-                    if (!vedtak) {
-                        return null;
-                    }
-
-                    return (
-                        <div key={s.søknad.id}>
-                            <Panel border className={styles.søknad}>
-                                <div className={styles.info}>
-                                    <div>
-                                        <Heading level="3" size="small" spacing>
-                                            {props.intl.formatMessage({ id: 'søknad.typeSøknad' })}
-                                        </Heading>
-                                        <div className={styles.dato}>
-                                            <Label>{`${props.intl.formatMessage({ id: 'søknad.mottatt' })}: `}</Label>
-                                            <BodyShort>{søknadMottatt(s.søknad, props.intl)}</BodyShort>
-                                        </div>
-                                        <div className={styles.dato}>
-                                            <Label>
-                                                {`${props.intl.formatMessage({
-                                                    id: 'søknad.iverksattDato',
-                                                })}: `}
-                                            </Label>
-                                            <BodyShort>{props.intl.formatDate(s.iverksattDato)}</BodyShort>
-                                        </div>
-                                    </div>
-                                    <div>
-                                        <LinkAsButton
-                                            variant="secondary"
-                                            href={Routes.vedtaksoppsummering.createURL({
-                                                sakId: props.sak.id,
-                                                vedtakId: vedtak.id,
-                                            })}
-                                            size="small"
-                                        >
-                                            {props.intl.formatMessage({ id: 'behandling.seOppsummering' })}
-                                        </LinkAsButton>
-                                    </div>
-                                </div>
-                            </Panel>
-                        </div>
-                    );
-                })}
-            </ol>
-        </div>
-    );
-};
-
-const StartSøknadsbehandlingKnapper = (props: { sakId: string; søknadId: string; intl: IntlShape }) => {
-    const history = useHistory();
-    const [behandlingStatus, startBehandling] = useAsyncActionCreator(sakSlice.startBehandling);
-
-    return (
-        <div>
-            <AvsluttOgStartFortsettButtons
-                sakId={props.sakId}
-                behandlingsId={props.søknadId}
-                primaryButtonTekst={props.intl.formatMessage({
-                    id: 'behandling.startBehandling',
-                })}
-                usePrimaryAsButton={{
-                    onClick: () =>
-                        startBehandling(
-                            {
-                                sakId: props.sakId,
-                                søknadId: props.søknadId,
-                            },
-                            (response) => {
-                                history.push(
-                                    Routes.saksbehandlingVilkårsvurdering.createURL({
-                                        sakId: props.sakId,
-                                        behandlingId: response.id,
-                                    })
-                                );
-                            }
-                        ),
-                    status: behandlingStatus,
-                }}
-                intl={props.intl}
-            />
-
-            {RemoteData.isFailure(behandlingStatus) && (
-                <div className={styles.feil}>
-                    <ApiErrorAlert error={behandlingStatus.error} />
-                </div>
-            )}
-        </div>
-    );
-};
-
-const SøknadsbehandlingStartetKnapper = (props: {
-    behandling: Behandling;
-    sakId: string;
-    søknadId: string;
-    intl: IntlShape;
-}) => {
-    const user = useUserContext();
-    const { behandling } = props;
-
-    return (
-        <div>
-            {erTilAttestering(behandling) && (!user.isAttestant || user.navIdent === behandling.saksbehandler) && (
-                <div className={styles.ikonContainer}>
-                    <Ikon className={styles.ikon} kind="info-sirkel-fyll" width={'24px'} />
-                    <p>
-                        {props.intl.formatMessage({
-                            id: 'attestering.tilAttestering',
-                        })}
-                    </p>
-                </div>
-            )}
-
-            <div className={styles.søknadsbehandlingKnapper}>
-                {erTilAttestering(behandling) && user.isAttestant && user.navIdent !== behandling.saksbehandler ? (
-                    <LinkAsButton
-                        variant="secondary"
-                        size="small"
-                        href={Routes.attesterSøknadsbehandling.createURL({
-                            sakId: props.sakId,
-                            behandlingId: behandling.id,
-                        })}
-                    >
-                        {props.intl.formatMessage({
-                            id: 'attestering.attester',
-                        })}
-                    </LinkAsButton>
-                ) : (
-                    !erTilAttestering(behandling) &&
-                    !erIverksatt(behandling) &&
-                    user.navIdent !== pipe(behandling.attesteringer ?? [], last, toNullable)?.attestant && (
-                        <AvsluttOgStartFortsettButtons
-                            sakId={props.sakId}
-                            behandlingsId={props.søknadId}
-                            primaryButtonTekst={props.intl.formatMessage({
-                                id: 'behandling.fortsettBehandling',
-                            })}
-                            usePrimaryAsLink={{
-                                url: Routes.saksbehandlingVilkårsvurdering.createURL({
-                                    sakId: props.sakId,
-                                    behandlingId: behandling.id,
-                                    vilkar: hentSisteVurdertSaksbehandlingssteg(behandling),
-                                }),
-                            }}
-                            intl={props.intl}
-                        />
-                    )
-                )}
-            </div>
-        </div>
-    );
-};
-
-const LukkedeSøknader = (props: { lukkedeSøknader: Søknad[]; intl: IntlShape }) => {
-    if (props.lukkedeSøknader.length === 0) return null;
-
-    return (
-        <div className={styles.søknadsContainer}>
-            <Heading level="2" size="medium" spacing>
-                {props.intl.formatMessage({
-                    id: 'søknad.lukkedeSøknader.tittel',
-                })}
-            </Heading>
-            <ol>
-                {props.lukkedeSøknader.map((søknad) => (
-                    <li key={søknad.id}>
-                        <Panel border className={styles.søknad}>
-                            <div className={styles.info}>
-                                <div>
-                                    <Heading level="3" size="small" spacing>
-                                        {props.intl.formatMessage({ id: 'søknad.typeSøknad' })}
-                                    </Heading>
-                                    <div className={styles.dato}>
-                                        <Label>{`${props.intl.formatMessage({ id: 'søknad.mottatt' })}: `}</Label>
-                                        <BodyShort>{søknadMottatt(søknad, props.intl)}</BodyShort>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className={styles.ikonContainer}>
-                                <Ikon className={styles.ikon} kind="feil-sirkel-fyll" width={'24px'} />
-                                <p className={styles.ikonTekst}>
-                                    {props.intl.formatMessage({
-                                        id: lukketBegrunnelseResourceId(søknad.lukket?.type),
-                                    })}
-                                </p>
-                            </div>
-                        </Panel>
-                    </li>
-                ))}
-            </ol>
-        </div>
-    );
-};
-
-const AvslåtteSøknader = (props: {
-    sak: Sak;
-    avslåtteSøknader: Array<{
-        iverksattDato: string | undefined;
-        søknadensBehandlingsId: string | undefined;
-        søknad: Søknad;
-    }>;
-    intl: IntlShape;
-}) => {
-    if (props.avslåtteSøknader.length === 0) return null;
-
-    return (
-        <div className={styles.søknadsContainer}>
-            <Heading level="2" size="medium" spacing>
-                {props.intl.formatMessage({
-                    id: 'søknad.avslåtteSøknader.tittel',
-                })}
-            </Heading>
-            <ol>
-                {props.avslåtteSøknader.map((s) => {
-                    if (!s.søknadensBehandlingsId) {
-                        return null;
-                    }
-                    const vedtak = props.sak.vedtak.find((v) => v.behandlingId === s.søknadensBehandlingsId);
-                    if (!vedtak) {
-                        return null;
-                    }
-
-                    return (
-                        <li key={s.søknad.id}>
-                            <Panel border className={styles.søknad}>
-                                <div className={styles.info}>
-                                    <div>
-                                        <Heading level="3" size="small" spacing>
-                                            {props.intl.formatMessage({ id: 'søknad.typeSøknad' })}
-                                        </Heading>
-                                        <div className={styles.dato}>
-                                            <Label>{`${props.intl.formatMessage({ id: 'søknad.mottatt' })}: `}</Label>
-                                            <BodyShort>{søknadMottatt(s.søknad, props.intl)}</BodyShort>
-                                        </div>
-                                        <div className={styles.dato}>
-                                            <Label>
-                                                {`${props.intl.formatMessage({
-                                                    id: 'søknad.iverksattDato',
-                                                })}: `}
-                                            </Label>
-                                            <BodyShort>{props.intl.formatDate(s.iverksattDato)}</BodyShort>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div>
-                                    <LinkAsButton
-                                        variant="secondary"
-                                        href={Routes.vedtaksoppsummering.createURL({
-                                            sakId: props.sak.id,
-                                            vedtakId: vedtak.id,
-                                        })}
-                                        size="small"
-                                    >
-                                        {props.intl.formatMessage({ id: 'revurdering.oppsummering' })}
-                                    </LinkAsButton>
-                                </div>
-                            </Panel>
-                        </li>
-                    );
-                })}
-            </ol>
-        </div>
-    );
-};
-
-const AvsluttOgStartFortsettButtons = (props: {
+export const AvsluttOgStartFortsettButtons = (props: {
     sakId: string;
     behandlingsId: string;
     primaryButtonTekst: string;
@@ -699,21 +143,24 @@ const AvsluttOgStartFortsettButtons = (props: {
         onClick: () => Promise<void>;
         status: ApiResult<Behandling, string>;
     };
+    hideSecondaryButton?: boolean;
 }) => {
     return (
         <div className={styles.avsluttOgStartFortsettKnapperContainer}>
-            <LinkAsButton
-                variant="secondary"
-                size="small"
-                href={Routes.avsluttBehandling.createURL({
-                    sakId: props.sakId,
-                    id: props.behandlingsId,
-                })}
-            >
-                {props.intl.formatMessage({
-                    id: 'behandling.avsluttBehandling',
-                })}
-            </LinkAsButton>
+            {!props.hideSecondaryButton && (
+                <LinkAsButton
+                    variant="secondary"
+                    size="small"
+                    href={Routes.avsluttBehandling.createURL({
+                        sakId: props.sakId,
+                        id: props.behandlingsId,
+                    })}
+                >
+                    {props.intl.formatMessage({
+                        id: 'behandling.avsluttBehandling',
+                    })}
+                </LinkAsButton>
+            )}
             {props.usePrimaryAsButton ? (
                 <Button size="small" onClick={props.usePrimaryAsButton.onClick}>
                     {props.primaryButtonTekst}
