+++ src/pages/saksbehandling/sakintro/sakintro.module.less
@@ -80,6 +80,10 @@
                 display: flex;
             }
 
+            .informasjonsTekst {
+                display: flex;
+            }
+
             .underkjenteAttesteringerContainer {
                 margin-top: @spacing-xs;
             }
